refactor(SD-ARCH-HOTSPOT-CHECKIN-001): extract resolveCheckin ladder into lib/checkin step pipeline
resolveCheckin (scripts/worker-checkin.cjs) shrinks from ~590 lines to 33: rungs 1+2
stay inline as ctx assembly; every other rung moved VERBATIM into 14 step modules under
lib/checkin/steps/ (ordered registry in index.cjs), driven by lib/checkin/pipeline.cjs
runSteps (no runner try/catch — step throws propagate exactly as before). Steps receive
dependencies via ctx.helpers (CHECKIN_HELPERS) so they never require worker-checkin.cjs
back (no cycle). module.exports and CLI wiring byte-identical; only permitted edit inside
moved blocks is local-var -> ctx.* substitution (callsign/mySd/sessionMetadata/base).
New tests/unit/checkin-pipeline.test.js pins registry order, step shape, short-circuit +
applies() skip semantics, and throw propagation.

Coordinator-authorized pin retarget: 4 source-location pin tests (checkin-merged-claim-pool,
worker-checkin-{fleet-critical,newest,ranked}-window) readFileSync'd worker-checkin.cjs and
regex-asserted the merged-pool wiring text inline in runCheckin — the rung moved verbatim to
lib/checkin/steps/merged-pool-self-claim.cjs, so only the readFileSync target + slice bounds
were retargeted; every regex assertion is byte-identical (all 15 pinned fragments verified
present in the step file). Full family green: 28 files / 373 tests (368 baseline + 5 new).


Fleet-Worker: Alpha
