+++ .worktree.json
@@ -1,7 +1,7 @@
 {
-  "sdKey": "SD-FDBK-ENH-APA-CALIBRATION-FIXTURE-001",
-  "expectedBranch": "feat/SD-FDBK-ENH-APA-CALIBRATION-FIXTURE-001",
-  "createdAt": "2026-07-08T10:50:25.391Z",
+  "sdKey": "SD-ARCH-HOTSPOT-CHECKIN-001",
+  "expectedBranch": "feat/SD-ARCH-HOTSPOT-CHECKIN-001",
+  "createdAt": "2026-07-09T22:12:52.800Z",
   "repoRoot": "C:\\Users\\rickf\\Projects\\_EHG\\EHG_Engineer",
   "baseRef": "origin/main"
 }
