+++ scripts/worker-checkin.cjs
@@ -67,6 +67,11 @@ const { NATO, COLORS, nextAvailable, isTestSessionId, tierRankOf, pickCallsignFo
 // claim must not be stolen while the prior claimant still has real, unmerged work.
 const { isSessionAlive } = require('../lib/fleet/session-liveness.cjs');
 const { hasWip } = require('../lib/claim/wip-detector.cjs');
+// SD-ARCH-HOTSPOT-CHECKIN-001: resolveCheckin's rung ladder is now an explicit step pipeline.
+// Steps live in lib/checkin/steps/ (VERBATIM moves of the former inline rungs) and receive their
+// dependencies via ctx.helpers (CHECKIN_HELPERS below) — steps never require this file (circular).
+const { runSteps } = require('../lib/checkin/pipeline.cjs');
+const CHECKIN_STEPS = require('../lib/checkin/steps/index.cjs');
 
 const ROLL_CALL_TTL_MS = 60 * 60 * 1000;     // availability row lives 1h
 const ROLL_CALL_DEDUP_MS = 5 * 60 * 1000;    // don't re-register within 5m (idempotency)
@@ -1323,576 +1328,19 @@ async function resolveCheckin(sb, sessionId, { getCoordinator = getActiveCoordin
     }
   } catch { /* fail-open */ }
 
-  // 2c. SD-LEO-INFRA-AUTO-TIERING-ACTIVATION-001-B (FR-3): merge --model/--effort into
-  // sessionMetadata BEFORE it is used below (tierCtx.worker_tier_rank at the resolveWorkerTierRank
-  // call further down reads sessionMetadata, so this call's own fresh values must land here first).
-  // Fail-open: a persist error never blocks the check-in itself.
-  try {
-    const { changed } = mergeCheckinModelEffort(sessionMetadata, { model: cliModel, effort: cliEffort });
-    if (changed) {
-      // QF-20260703-314: re-read metadata FRESH immediately before writing instead of persisting
-      // the whole step-2 snapshot merged in-memory -- a concurrent writer (e.g.
-      // assign-fleet-identities.cjs setting fleet_identity) landing between step 2's read and this
-      // write was otherwise silently clobbered by this whole-object update (mechanism (c) of the
-      // callsign-collision chain, backlog cfea31a7: worker 3c40949b lost Charlie/red within 9min).
-      const { data: freshRow } = await sb.from('claude_sessions').select('metadata').eq('session_id', sessionId).maybeSingle();
-      const freshBase = (freshRow && freshRow.metadata) || sessionMetadata || {};
-      const { metadata: freshMerged } = mergeCheckinModelEffort(freshBase, { model: cliModel, effort: cliEffort });
-      sessionMetadata = freshMerged;
-      await sb.from('claude_sessions').update({ metadata: freshMerged }).eq('session_id', sessionId);
-    }
-  } catch { /* fail-open: check-in proceeds with whatever metadata was already read */ }
-
-  // 2c-2. SD-LEO-INFRA-CLAIM-BOUNDARY-PRE-001 FR-4: self-clear an uncleared probe quarantine.
-  // This checkin executing IS the resumed-tool-activity proof (its Bash call fired PostToolUse).
-  // Must run BEFORE the 5.9 isSelfClaimDisabled gate so a recovered window re-enters self-claim
-  // in this same pass instead of idling one extra cycle.
-  sessionMetadata = await selfClearQuarantine(sb, sessionId, sessionMetadata);
-
-  // 2b. FR-2: re-hydrate callsign from the durable SET_IDENTITY row if metadata lost it (survives
-  // release/sweep). Runs BEFORE registerRollCall so the re-hydrated callsign flows into the roll-call
-  // row + base.callsign in the SAME response.
-  if (!callsign) {
-    const rehydrated = await rehydrateCallsign(sb, sessionId, sessionMetadata);
-    if (rehydrated) callsign = rehydrated;
-  }
-
-  // 3. register availability (idempotent)
-  const rollCall = await registerRollCall(sb, { sessionId, coordinatorId, callsign, mySd });
-
-  const base = { ok: true, callsign, coordinator: coordinatorId, roll_call_id: rollCall.id, two_way: process.env.COORDINATOR_TWOWAY_V2 === 'on' };
-
-  // SD-LEO-INFRA-WORKER-WINDDOWN-SURVEY-001 (b): surface the prior wind-down reason captured by the
-  // Stop hook (claude_sessions.metadata.wind_down) so the /checkin skill can render "you previously
-  // stopped because X — confirm/correct" and the worker can correct the inferred reason at re-engage.
-  base.prior_wind_down = (sessionMetadata && sessionMetadata.wind_down) ? sessionMetadata.wind_down : null;
-
-  // FR-1/FR-3: surface UNCONSUMED coordinator->worker push as coordinator_messages[] on the `base`
-  // object so EVERY return path (resume / idle / self_claimed / self_claimed_qf) carries it — a busy
-  // claim-holder AND an idle worker both see coordinator coaching. Non-draining + bounded (see fn).
-  base.coordinator_messages = await surfaceCoordinatorMessages(sb, sessionId, { role: sessionRole });
-
-  // 4. already working -> resume. A self-claimed quick-fix lands in claude_sessions.sd_key
-  // too (claim_sd writes it for QF-% ids), so a QF claim must resume into the /quick-fix
-  // workflow — NOT sd-start, which is SD-only (QFs have no worktree / LEAD-PLAN-EXEC).
-  if (mySd) {
-    const isQf = /^QF-/.test(mySd);
-    // FR-2 (SD-LEO-INFRA-CLAIM-LIFECYCLE-HARDENING-002): a stale claude_sessions.sd_key pointing at
-    // a TERMINAL/parked SD (completed/cancelled/deferred) would loop action=resume forever. Before
-    // resuming, verify the SD is still resumable; if terminal, self-heal (CAS-guarded to THIS session)
-    // and fall through to self-claim. pending_approval is NOT terminal — it resumes (to run
-    // LEAD-FINAL-APPROVAL) and recoverStrandedFinal (step 5.7) owns the cleared-claim variant.
-    // Fail-open: any query error preserves today's resume.
-    let staleTerminal = false;
-    // SD-LEO-FEAT-WORKER-CHECKIN-SELF-001 (FR-1): a HARD-DELETED claimed SD (row gone) used to fall
-    // through every guard and loop action=resume on a ghost forever (sd-start then exits SD-not-found
-    // = permanent strand). Track it separately and self-heal it like a stale-terminal claim.
-    let staleDeleted = false;
-    if (!isQf) {
-      try {
-        const { data: sdRow, error: sdErr } = await sb.from('strategic_directives_v2').select('status').eq('sd_key', mySd).maybeSingle();
-        if (sdErr) throw sdErr; // read error -> fail-open (preserve resume)
-        if (sdRow) {
-          if (['completed', 'cancelled', 'deferred'].includes(sdRow.status)) staleTerminal = true;
-        } else if (await confirmRowGone(sb, 'strategic_directives_v2', 'sd_key', mySd)) {
-          // FR-2: only after a CONFIRMING re-read also finds the row absent (never release on a single
-          // transient/eventual-consistency null).
-          staleDeleted = true;
-        }
-      } catch { /* fail-open: leave flags false -> resume preserved */ }
-    } else {
-      // Quick-fix QF-20260612-113: QF claims land in claude_sessions.sd_key too, but the
-      // terminal self-heal above was SD-only — a completed/cancelled/escalated QF looped
-      // action=resume forever. Apply the same check against quick_fixes.status.
-      try {
-        const { data: qfRow, error: qfErr } = await sb.from('quick_fixes').select('status').eq('id', mySd).maybeSingle();
-        if (qfErr) throw qfErr;
-        if (qfRow) {
-          if (['completed', 'cancelled', 'escalated'].includes(qfRow.status)) staleTerminal = true;
-        } else if (await confirmRowGone(sb, 'quick_fixes', 'id', mySd)) {
-          staleDeleted = true; // FR-1/FR-2: a hard-deleted quick-fix, confirmed absent
-        }
-      } catch { /* fail-open: leave flags false -> resume preserved */ }
-    }
-    if (staleTerminal || staleDeleted) {
-      await selfHealStaleClaim(sb, sessionId, mySd);
-      base.self_healed_stale_claim = mySd;
-      if (staleDeleted) base.self_healed_deleted_claim = mySd; // FR-1: distinguish deleted from terminal
-      mySd = null; // fall through to assignment / self-claim below
-    } else {
-      // SD-FDBK-FIX-WORKER-CHECK-SURFACES-001 (seam 1): a claim-holding worker used to
-      // short-circuit to resume BEFORE the step-5 WORK_ASSIGNMENT pull, so a coordinator
-      // directive targeting a BUSY worker was never read on check-in (witnessed: an assignment to
-      // Alpha stayed UNREAD across full cycles). Peek for a WORK_ASSIGNMENT targeting THIS session
-      // for a DIFFERENT SD and SURFACE it on the resume result so the worker sees it now. We do NOT
-      // drain read_at and do NOT auto-switch the claim — never-strand (CLAUDE.md rule 7a): the
-      // worker finishes / explicitly hands off current work, then actions the assignment (genuine
-      // claim_sd/ackMessage stamps read_at then). Surfacing persists (seam 2 re-surfaces on poll)
-      // until actioned. Fail-open: any error preserves today's plain resume.
-      let pendingAssignment = null;
-      try {
-        // QF-20260703-476: unackedOnly (acknowledged_at IS NULL), not unreadOnly -- a row whose
-        // read_at got stamped by some other delivery path but was never genuinely actioned must
-        // still surface here, bounded to a recency window so an ancient row isn't resurrected.
-        const msgs = await ws.getMessagesForSession(sb, sessionId, {
-          unackedOnly: true,
-          sinceIso: new Date(Date.now() - ASSIGNMENT_RECENCY_WINDOW_MS).toISOString(),
-        });
-        // Only a row carrying a STRUCTURED directed field (assigned_sd/sd_key) is a real redirect;
-        // selecting on extractDirectedSd also skips past the generic sweep advisory if a directed
-        // row exists beneath it (finding #3: subtype-blind newest-wins .find() could otherwise mask
-        // a genuine redirect). getMessagesForSession returns created_at DESC, so this is newest-first.
-        const wa = (msgs || []).find(m => m.message_type === 'WORK_ASSIGNMENT' && extractDirectedSd(m));
-        if (wa) {
-          const waSd = extractDirectedSd(wa);
-          // QF-20260705-858: a preempt-flagged WA (coordinator dispatch stamps payload.preempt=true
-          // on a chairman-critical redirect) must surface with an act-now framing, distinct from the
-          // routine "finish current work first" note below -- the resume short-circuit otherwise
-          // structurally delays a preempt until the current SD fully completes. Non-preempt WAs are
-          // BYTE-IDENTICAL to prior behavior (no new field, no message change) — only the preempt
-          // branch changes.
-          if (waSd && waSd !== mySd) pendingAssignment = { sd: waSd, message_id: wa.id, preempt: wa.payload?.preempt === true };
-        }
-      } catch { /* fail-open: no pending-assignment surfacing */ }
-      const resumeMsg = isQf
-        ? `Already claiming quick-fix ${mySd}; resume it: node scripts/read-quick-fix.js ${mySd}, then run the /quick-fix workflow (do NOT run sd-start.js for a QF).`
-        : `Already claiming ${mySd}; resume work (run sd-start to (re)attach the worktree).`;
-      let resumeMessage = resumeMsg;
-      if (pendingAssignment?.preempt) {
-        resumeMessage = `${resumeMsg} ⚠ PREEMPT: coordinator has redirected you to ${pendingAssignment.sd} (chairman-critical). Consider releasing ${mySd} cleanly (worker owns clean suspension — never auto-released) and claiming ${pendingAssignment.sd} now, rather than finishing ${mySd} first.`;
-      } else if (pendingAssignment) {
-        resumeMessage = `${resumeMsg} NOTE: coordinator WORK_ASSIGNMENT pending for ${pendingAssignment.sd} — finish/hand off ${mySd} first, then claim it (never drop an in-progress SD).`;
-      }
-      return { ...base, action: 'resume', sd: mySd,
-        ...(pendingAssignment ? { pending_work_assignment: pendingAssignment } : {}),
-        message: resumeMessage };
-    }
-  }
-
-  // 4.5 ACQUISITION GUARD (SD-FDBK-FIX-SELF-ONLY-AUTHORIZATION-001, feedback a159d1ec):
-  // propose-only sessions (metadata.non_fleet=true / role=adam, CONST-002) must NEVER
-  // acquire a build claim. The claim-validity gate's CHECK 1.5 only fires at sd-start/
-  // handoff — but a propose-only session is the LEAST likely to ever reach a handoff, so
-  // that tripwire never trips and an Adam session could self-claim here and starve real
-  // workers (claim_sd surfaces it as a live foreign holder). Short-circuit to idle BEFORE
-  // every acquisition tier (assignment, recoverStrandedFinal, adoptOrphanInProgress,
-  // self-claim, QF). The resume of a pre-existing claim above is intentionally NOT blocked
-  // (legacy state; the gate still blocks its handoffs). Fail-safe: only an explicit
-  // non_fleet/adam triggers. A follow-up SD adds the symmetric guard inside the claim_sd
-  // RPC (covers qf-start / sweep / reacquire callers too).
-  if (isBuildForbiddenSession(sessionMetadata)) {
-    return { ...base, action: 'idle', recommended_wakeup_seconds: 1200,
-      message: 'Propose-only session (non_fleet / role=adam): build claims are forbidden per CONST-002 — not self-claiming. Adam proposes work via the decision queue.' };
-  }
-
-  // 5. pending WORK_ASSIGNMENT -> claim via claim_sd RPC
-  // QF-20260703-476: unackedOnly, not unreadOnly -- see ASSIGNMENT_RECENCY_WINDOW_MS above. A
-  // consumed-but-unactioned row (read_at set, acknowledged_at NULL, no claim recorded) must still
-  // reach the claim step instead of being permanently hidden by an unreadOnly filter.
-  let assignment = null;
-  const assignmentSinceIso = new Date(Date.now() - ASSIGNMENT_RECENCY_WINDOW_MS).toISOString();
-  try {
-    const msgs = await ws.getMessagesForSession(sb, sessionId, { unackedOnly: true, sinceIso: assignmentSinceIso });
-    // QF-20260705-914: informational completion nudges are never claimable assignments.
-    assignment = (msgs || []).find(m => m.message_type === 'WORK_ASSIGNMENT' && !isInformationalNudge(m));
-    if (!assignment) {
-      // QF-20260703-806: the unacked pull can miss a row whose acknowledged_at got stamped by a
-      // path OTHER than a genuine claim outcome (the ack-before-claim race). Claim outcome, not
-      // ack, is the terminal state -- widen to the same bounded window WITHOUT the ack filter.
-      // The terminal/ineligible/tryClaim checks below already re-verify the target SD's LIVE
-      // state and are idempotent, so resurrecting an already-genuinely-resolved row is harmless.
-      const wider = await ws.getMessagesForSession(sb, sessionId, { sinceIso: assignmentSinceIso });
-      assignment = (wider || []).find(m => m.message_type === 'WORK_ASSIGNMENT' && !isInformationalNudge(m));
-    }
-  } catch { /* fail-open */ }
-  if (assignment) {
-    const sdKey = extractSdFromAssignment(assignment);
-    if (sdKey) {
-      // SD-LEO-FIX-CLAIM-RPC-TERMINAL-001: purge a STALE assignment whose target SD reached a
-      // terminal status (completed/cancelled/deferred) AFTER the assignment was created — the
-      // in_progress->terminal race. claim_sd now refuses terminal claims, but a never-ACKed
-      // assignment would re-fire every tick (the "retried on every tick" symptom), so ACK it
-      // here and fall through to self-claim.
-      // QF-20260704-602: this strategic_directives_v2 lookup is intentionally SD-only. A
-      // directed QF key (sdKey starting 'QF-') always misses here (assignedSdRow stays null),
-      // which correctly skips terminalStatus/ineligibleReason below (both SD-specific concerns
-      // with no QF equivalent yet) and falls straight to tryClaim() — the claim_sd RPC is
-      // already QF-aware (p_sd_id LIKE 'QF-%' branches to the quick_fixes table) and does not
-      // need this fetch to succeed.
-      let terminalStatus = null;
-      let assignedSdRow = null;
-      let assignedSdFetchFailed = false;
-      try {
-        const { data: tgt, error: tgtErr } = await sb.from('strategic_directives_v2')
-          .select('status, sd_type, sd_key, metadata, target_application').eq('sd_key', sdKey).maybeSingle();
-        // QF-20260703-151: a query ERROR is distinct from a genuine not-found and must NOT be
-        // silently discarded — the prior code destructured only `data`, so a failed fetch left
-        // assignedSdRow=null, which the ineligibleReason ternary below then treated as "nothing
-        // to check" (fail-open), admitting an orchestrator-parent / repo-mismatched SD straight
-        // through to tryClaim (live-hit: SD-EHG-PRODUCT-UIUX-REMEDIATION-001, audit flag c71c3a54).
-        if (tgtErr) {
-          assignedSdFetchFailed = true;
-        } else {
-          assignedSdRow = tgt || null;
-          if (tgt && ['completed', 'cancelled', 'deferred'].includes(tgt.status)) terminalStatus = tgt.status;
-        }
-      } catch { assignedSdFetchFailed = true; }
-      // QF-20260705-429 (residual of QF-20260705-460): directed QF assignments must honor
-      // quick_fixes.not_before (durable time-gated defer). The SD lookup above intentionally
-      // misses for QF- keys and the claim_sd RPC never reads not_before, so without this gate
-      // a directed assignment claims a deferred QF hours early (specimen: QF-20260704-348,
-      // ~2.6h before its gate). Deferral is TRANSIENT: on gate, do NOT ack — the assignment
-      // stays live and succeeds once not_before passes. Read error = FAIL CLOSED via
-      // assignedSdFetchFailed, mirroring the QF-20260703-151 semantics above.
-      // QF-20260705-115: same query also carries `status`, so a directed assignment whose QF
-      // already reached a terminal status is purged HERE instead of relying on claim_sd's RPC
-      // rejection to land in TERMINAL_CLAIM_ERRORS (it doesn't always -- live specimen:
-      // QF-20260705-436 resurfaced 5+ checkins post-completion). Reuses `terminalStatus` so it
-      // takes the SAME ack+stale_assignment_purged branch as the SD terminal case below.
-      // Terminal set mirrors the existing quick_fixes staleness check at line ~1386.
-      let qfDeferredUntil = null;
-      if (!assignedSdFetchFailed && /^QF-/.test(sdKey)) {
-        try {
-          const { data: qfRow, error: qfErr } = await sb.from('quick_fixes')
-            .select('status, not_before').eq('id', sdKey).maybeSingle();
-          if (qfErr) {
-            assignedSdFetchFailed = true;
-          } else if (qfRow && ['completed', 'cancelled', 'escalated'].includes(qfRow.status)) {
-            terminalStatus = qfRow.status;
-          } else {
-            const nb = qfRow && qfRow.not_before ? Date.parse(qfRow.not_before) : NaN;
-            if (Number.isFinite(nb) && nb > Date.now()) qfDeferredUntil = qfRow.not_before;
-          }
-        } catch { assignedSdFetchFailed = true; }
-      }
-      // QF-20260703-091 (RCA-confirmed): mirror the self-claim paths' shared fitness/premise gate
-      // (classifyDispatchIneligibility, incl. the repo-match axis) onto this DIRECTED-assignment
-      // path too. Previously only sd-start.js caught a repo-mismatched directed assignment, AFTER
-      // the claim had already churned (coordinator re-dispatch -> checkin claims unconditionally ->
-      // sd-start releases -> repeats every tick). No tierCtx: an explicit coordinator directive
-      // should not be second-guessed by the WORK-DOWN-NEVER-UP self-claim preference, only the hard
-      // fitness/premise axes (repo mismatch, terminal/superseded premise, orchestrator parent, etc).
-      const ineligibleReason = (!terminalStatus && assignedSdRow)
-        ? classifyDispatchIneligibility(assignedSdRow, { cwd: process.cwd() })
-        : null;
-      if (assignedSdFetchFailed) {
-        // QF-20260703-151: FAIL CLOSED — could not confirm fitness for this assignment. Never
-        // ack (the assignment stays live for a retry once the transient fetch issue clears) and
-        // never fall through to tryClaim on an unconfirmed target.
-        base.assignment_claim_error = 'fitness_check_query_failed';
-      } else if (terminalStatus) {
-        await ackMessage(sb, assignment.id, { role: sessionRole, kind: assignment.payload?.kind, messageType: assignment.message_type });
-        base.stale_assignment_purged = { sd: sdKey, status: terminalStatus };
-      } else if (ineligibleReason) {
-        await ackMessage(sb, assignment.id, { role: sessionRole, kind: assignment.payload?.kind, messageType: assignment.message_type });
-        base.assignment_ineligible_purged = { sd: sdKey, reason: ineligibleReason };
-      } else if (qfDeferredUntil) {
-        // No ack: transient — re-attempted next tick; claims automatically once the gate passes.
-        base.assignment_deferred_not_before = { qf: sdKey, not_before: qfDeferredUntil };
-      } else {
-        const claimed = await tryClaim(sb, sdKey, sessionId);
-        if (claimed.ok) {
-          await ackMessage(sb, assignment.id, { role: sessionRole, kind: assignment.payload?.kind, messageType: assignment.message_type });
-          // SD-MAN-INFRA-MEDIUM-EFFORT-HARDENING-001 (FR-5): surface the coordinator's
-          // ADVISORY effort recommendation so the worker banner can render it.
-          const effortRec = assignment.payload?.effort_recommendation || null;
-          return { ...base, action: 'claimed_assignment', sd: sdKey,
-            ...(effortRec ? { effort_recommendation: effortRec, effort_recommendation_reason: assignment.payload?.effort_recommendation_reason || null } : {}),
-            message: `Claimed assigned ${sdKey} via claim_sd.${effortRec ? ` Recommended effort: ${effortRec} (advisory).` : ''} Run: node scripts/sd-start.js ${sdKey}. ${antiWinddownDirective(base.belt_ranked_claimable)}` };
-        }
-        // QF-20260703-780: a terminal-class RPC verdict means this assignment can NEVER
-        // succeed (unlike e.g. claimed_by_live_peer, which may resolve next tick) -- ack it
-        // now so it stops being re-selected, mirroring the stale_assignment_purged /
-        // assignment_ineligible_purged branches above. Distinct breadcrumb so callers can
-        // tell "permanently resolved" apart from assignment_claim_error's "retryable" meaning.
-        if (TERMINAL_CLAIM_ERRORS.has(claimed.error)) {
-          await ackMessage(sb, assignment.id, { role: sessionRole, kind: assignment.payload?.kind, messageType: assignment.message_type });
-          base.assignment_claim_terminal_purged = { sd: sdKey, error: claimed.error };
-        } else {
-          // could not claim the assigned SD -> fall through to self-claim
-          base.assignment_claim_error = claimed.error;
-        }
-      }
-    }
-  }
-
-  // 5.7 SD-FDBK-FIX-RECURRING-2ND-OCCURRENCE-001: recover a STRANDED pending_approval/LEAD_FINAL SD
-  //      (claim cleared, one handoff from shipped) BEFORE self-claiming new work — finishing a
-  //      near-shipped SD beats starting fresh. Re-claiming lets a worker run LEAD-FINAL-APPROVAL with
-  //      a valid matching claim (passing the claim-validity gate the coordinator-from-main path fails).
-  const recovered = await recoverStrandedFinal(sb, sessionId, base);
-  if (recovered) return recovered;
-
-  // 5.8 SD-FDBK-INFRA-ORPHAN-ADOPTION-WORKER-001: adopt an ORPHANED in_progress SD (zero active
-  //     claims, session reaped mid-build) BEFORE self-claiming new work — finishing a
-  //     partially-built SD beats starting fresh, but a one-handoff-from-shipped stranded final
-  //     (5.7) still wins over a mid-build orphan.
-  const adopted = await adoptOrphanInProgress(sb, sessionId, base);
-  if (adopted) return adopted;
-
-  // 5.9 SD-REFILL-00BCYOYW: session-level self-claim opt-out. Placed AFTER the recovery tiers (so a
-  //     winding-down worker still finishes its own stranded/orphaned build work — that is not churn)
-  //     and BEFORE step-6, so it skips ONLY the self_claim-from-sd:next path (step-6 + selfClaimQuickFix).
-  //     roll_call, resume, directed WORK_ASSIGNMENT claiming and recovery all run before this point and
-  //     are unaffected. Narrower sibling of the 4.5 isBuildForbiddenSession guard (which blocks ALL
-  //     acquisition incl. directed work); this blocks only self-initiated claims.
-  if (isSelfClaimDisabled(sessionMetadata)) {
-    return { ...base, action: 'idle', recommended_wakeup_seconds: 1200,
-      message: 'Session self-claim disabled (metadata.self_claim=false / availability=idle_only / coordinator_stand_down=true): skipping self-claim from sd:next to avoid grab-release churn that blocks fresh-session pickup of reserved SDs. roll_call, directed WORK_ASSIGNMENTs and own stranded/orphan recovery still honored. Clear the flag to resume self-claim.' };
-  }
-  // SD-LEO-INFRA-SELF-CLAIM-STANDDOWN-HONOR-001 FR-2: a coordinator/global fleet stand-down halts the
-  // whole fleet's self-claim (overnight-reduction / keeper-only / venture-parked / hard-halt). roll_call,
-  // resume, directed WORK_ASSIGNMENT and recovery all ran above; only the sd:next self-claim is skipped.
-  if (await isGlobalStandDownActive(sb)) {
-    return { ...base, action: 'idle', recommended_wakeup_seconds: 1200,
-      message: 'Global/fleet stand-down active (system_settings FLEET_STAND_DOWN / HARD_HALT_STATUS enabled): skipping self-claim from sd:next. roll_call, directed WORK_ASSIGNMENTs and own stranded/orphan recovery still honored. Clear the system_settings stand-down to resume fleet self-claim.' };
-  }
-
-  // 5.5 SD-FDBK-INFRA-AUTO-MAINTAIN-EXECUTION-001: ensure an active execution baseline exists
-  //      BEFORE reading v_sd_next_candidates. With zero active baseline the view returns 0 rows
-  //      and self-claim silently idles with a full queue. Fail-open: a failure here degrades to
-  //      today's behavior (read returns [] -> idle), never an error action.
-  try { await ensureActiveBaseline(sb); } catch { /* fail-open: never block the checkin */ }
-
-  // 5.95 QF-20260704-244 (leg 3): a CRITICAL open QF, aged past the directed-dispatch grace
-  // window, outranks SD self-claim. Fenced to prevent reverse SD-belt starvation: only
-  // 'critical' jumps, and at most ONE consecutive jump per worker -- if the LAST pull was a
-  // jump (metadata.last_claim_was_qf_jump), this pull consumes/clears that flag and falls
-  // through to the normal SD-first order instead of jumping again immediately.
-  if (sessionMetadata?.last_claim_was_qf_jump === true) {
-    try {
-      await sb.from('claude_sessions')
-        .update({ metadata: { ...sessionMetadata, last_claim_was_qf_jump: false } })
-        .eq('session_id', sessionId);
-    } catch { /* fail-open: worst case is one extra consecutive jump */ }
-  } else {
-    try {
-      const { data: criticalQfs } = await sb
-        .from('quick_fixes')
-        .select('id, status, pr_url, commit_sha, created_at, routing_tier, title, severity, not_before')
-        .eq('status', 'open')
-        .eq('severity', 'critical')
-        .is('pr_url', null)
-        .is('commit_sha', null)
-        .order('created_at', { ascending: true })
-        .limit(QF_CANDIDATE_LIMIT);
-      const nowMs = Date.now();
-      for (const qf of (criticalQfs || [])) {
-        if (!isCriticalQfJumpEligible(qf, nowMs)) continue;
-        const claimed = await tryClaim(sb, qf.id, sessionId);
-        if (claimed.ok) {
-          try {
-            await sb.from('claude_sessions')
-              .update({ metadata: { ...sessionMetadata, last_claim_was_qf_jump: true } })
-              .eq('session_id', sessionId);
-          } catch { /* fail-open: worst case the one-consecutive bound doesn't hold once */ }
-          return {
-            ...base,
-            action: 'self_claimed_qf',
-            qf: qf.id,
-            message: `Self-claimed CRITICAL quick-fix ${qf.id} (priority jump ahead of SD self-claim, QF-20260704-244). Load it: node scripts/read-quick-fix.js ${qf.id} — then run the /quick-fix workflow (implement <=50 LOC on branch qf/${qf.id}, run tests, then node scripts/complete-quick-fix.js ${qf.id}). Do NOT run sd-start.js for a QF. On completion, re-run /checkin.`,
-          };
-        }
-      }
-    } catch { /* fail-open: proceed to normal SD-first order */ }
-  }
-
-  // 6. self-claim from ONE merged SD pool: baselined sd:next candidates + claimable
-  //    UN-BASELINED drafts, rank-sorted TOGETHER (QF-20260610-986, feedback dc87039d).
-  //    The old sequential tiers (6 then 6.25) meant a coordinator dispatch_rank could
-  //    never lift a draft above ANY baselined candidate — a rank-0 critical draft was
-  //    skipped for rank-5 baselined mediums. Merging preserves the no-rank precedence
-  //    (baselined entries listed first; orderByRankMap's stable sort keeps unranked
-  //    rows in input order) while fresh ranks reorder across both pools. Per-kind
-  //    eligibility guards and claim semantics unchanged; baselined entry wins a
-  //    same-SD dedup (both pools can surface one SD).
-  try {
-    const { data: cands } = await sb
-      .from('v_sd_next_candidates')
-      .select('sd_id, track, status, priority')
-      .limit(SELF_CLAIM_CANDIDATE_LIMIT);
-    let draftRows = [];
-    try { draftRows = await fetchDraftCandidates(sb); } catch { /* fail-open: drafts absent */ }
-    // SD-LEO-INFRA-SELF-CLAIM-WINDOW-NEWEST-FIT-DRAFT-001 (FR-1): a FOURTH source — the NEWEST-N drafts
-    // (created_at DESC), so a fresh non-fleet_critical fit-draft at age-position 11+ (outside the oldest-10
-    // fetchDraftCandidates window and not fleet_critical) is in the pool instead of starving. Fail-open,
-    // mirroring the draft/fleet_critical fetches.
-    let newestRows = [];
-    try { newestRows = await fetchNewestDraftCandidates(sb); } catch { /* fail-open: newest window absent */ }
-    // SD-LEO-INFRA-SELF-CLAIM-WINDOW-FLEET-CRITICAL-001 (FR-2): a THIRD source for fleet_critical SDs
-    // that sit OUTSIDE both windows above, so the downstream fleet_critical lift has them in the pool to
-    // reorder. Placed here, inside the step-6 try and downstream of ALL acquisition guards (4.5/5.7/5.8/5.9).
-    let fcRows = [];
-    try { fcRows = await fetchFleetCriticalCandidates(sb); } catch { /* fail-open: window-only behavior */ }
-    // SD-LEO-INFRA-GUARANTEE-CLAIMABLE-SD-RANKED-001-B (FR-1/FR-2): a FIFTH source for SDs the
-    // coordinator has ranked (metadata.dispatch_rank set) that sit OUTSIDE every window above, so a
-    // ranked-but-not-fleet_critical, middle-of-the-backlog SD is not wasted ranking work.
-    let rankedRows = [];
-    try { rankedRows = await fetchRankedCandidates(sb); } catch { /* fail-open: window-only behavior */ }
-
-    const seen = new Set();
-    const merged = [];
-    for (const c of (cands || [])) {
-      if (c.sd_id && !seen.has(c.sd_id)) { seen.add(c.sd_id); merged.push({ kind: 'baselined', key: c.sd_id, track: c.track }); }
-    }
-    for (const d of draftRows) {
-      if (d.sd_key && !seen.has(d.sd_key)) { seen.add(d.sd_key); merged.push({ kind: 'draft', key: d.sd_key, row: d }); }
-    }
-    // SD-LEO-INFRA-SELF-CLAIM-WINDOW-NEWEST-FIT-DRAFT-001 (FR-1/FR-3): union the NEWEST-N drafts as
-    // kind:'draft' rows (so each routes through tryClaimDraftCandidate -> the COMPLETE eligibility SSOT),
-    // deduped by the SAME seen-set so an SD already surfaced by the oldest-10 window keeps its existing
-    // entry (no double-count). Widens the POOL only; changes NO eligibility/ordering semantics.
-    for (const d of newestRows) {
-      if (d.sd_key && !seen.has(d.sd_key)) { seen.add(d.sd_key); merged.push({ kind: 'draft', key: d.sd_key, row: d }); }
-    }
-    // FR-2: union the fleet_critical source LAST so an SD already surfaced by the view/draft windows keeps
-    // its existing entry (dedup via the SAME seen-set). kind:'baselined' routes each injected entry through
-    // baselinedCandidateEligible -> the COMPLETE eligibility SSOT (classifyDispatchIneligibility incl. the
-    // WORK-DOWN-NEVER-UP tier axis + parentLeadPending + refillSourceIneligibility + draftDepsSatisfied),
-    // then isSdInFlight, then tryClaim — NO eligibility/claim bypass. sortByDispatchRank then lifts these
-    // (strict-boolean fleet_critical) to the front of the merged pool.
-    for (const f of fcRows) {
-      if (f.sd_key && !seen.has(f.sd_key)) { seen.add(f.sd_key); merged.push({ kind: 'baselined', key: f.sd_key }); }
-    }
-    // FR-2 (SD-LEO-INFRA-GUARANTEE-CLAIMABLE-SD-RANKED-001-B): union the ranked-direct source LAST so
-    // an SD already surfaced by any prior source keeps its existing entry (SAME seen-set dedup).
-    // kind:'baselined' routes each injected entry through baselinedCandidateEligible -> the COMPLETE
-    // eligibility SSOT, exactly like the fleet_critical union above — no eligibility/claim bypass.
-    for (const r of rankedRows) {
-      if (r.sd_key && !seen.has(r.sd_key)) { seen.add(r.sd_key); merged.push({ kind: 'baselined', key: r.sd_key }); }
-    }
-
-    // duty-6: honor the coordinator's fresh dispatch_rank across the WHOLE pool (fail-open).
-    const ranked = await sortByDispatchRank(sb, merged, (x) => x.key);
-    // FR-1 (anti-premature-winddown): expose the ranked claimable belt depth on EVERY result so the
-    // /checkin skill can render concrete available work — a worker about to wind down sees data, not a
-    // vibe. base is spread into the self_claimed / idle / QF results below.
-    base.belt_ranked_claimable = ranked.length;
-    // SD-LEO-INFRA-COMPLEXITY-TIERED-WORKER-ASSIGNMENT-001 (FR-3 + FR-5): resolve this worker's rung
-    // and whether tiering is active ONCE per checkin (both are per-run constants, not per-candidate),
-    // then thread into the shared classifier so a below-rung worker skips above-rung work. Fail-open:
-    // any fault leaves tierCtx empty => byte-identical pre-tiering behavior.
-    let tierCtx = {};
-    try {
-      tierCtx = {
-        worker_tier_rank: resolveWorkerTierRank({ metadata: sessionMetadata }),
-        tiering_active: await isTieringActive(sb),
-      };
-    } catch { /* fail-open: no tier ctx */ }
-    // SD-LEO-INFRA-AUTO-TIERING-ACTIVATION-001-E (FR-6): precompute the backlog verdict inputs ONCE
-    // per tick — lowerTierBacklog() itself is pure/sync (runs inside classifyDispatchIneligibility),
-    // but its two halves (claimable-by-tier, idle-by-tier) are DB-dependent, so fetchLowerTierBacklogData
-    // does the fetch here and the result is threaded into tierCtx.lower_tier_backlog_data. Only when
-    // tiering is active (mirrors every other tier-axis fetch in this function) — with < 2 live workers
-    // the whole tier axis, including this gate, must stay inert (degrade-to-1). Fail-open: a null
-    // return leaves the field unset, which is the classifier's documented byte-identical
-    // WORK-DOWN-ALWAYS fallback.
-    if (tierCtx.tiering_active === true) {
-      const backlogData = await fetchLowerTierBacklogData(sb);
-      if (backlogData) tierCtx.lower_tier_backlog_data = backlogData;
-      // QF-20260709-881: only a top-rung (fable) worker's downward claims are gated on this, so
-      // only fetch when relevant — avoids a config round-trip for every sub-top-rung checkin.
-      if (Number(tierCtx.worker_tier_rank) >= ladderTopRank()) {
-        tierCtx.fable_window_active = await fetchFableWindowActive(sb);
-      }
-    }
-    // QF-20260630-761: snapshot whether tiering is active so the idle message (below, outside this
-    // scope) only attributes a 0-claimable belt to TIER when tiering is actually on. With tiering off
-    // the 0 is non-tier ineligibility (orchestrator parents / clone trees / human-action / held).
-    base.belt_tiering_active = tierCtx.tiering_active === true;
-    // SD-LEO-INFRA-BELT-TIER-AWARE-CLAIMABILITY-001 (FR-2): belt_ranked_claimable above is the
-    // tier-AGNOSTIC ranked pool — a below-rung worker sees it non-zero even when every ranked SD is
-    // above its rung, then idles for hours on false "ranked" hope. Expose belt_claimable_at_my_tier:
-    // of the ranked pool, how many are base-eligible AND reachable at THIS worker's rung (shared
-    // tier-claimable rollup, reusing the gate). One batched fetch supplies the metadata the view-sourced
-    // baselined candidates lack. Fail-open to the agnostic count so a fault never under-reports.
-    base.belt_claimable_at_my_tier = base.belt_ranked_claimable;
-    try {
-      // QF-20260629-597: ALL ranked keys are sd_keys — baselined keys come from
-      // v_sd_next_candidates.sd_id which holds the sd_key STRING (not a UUID), and draft keys are sd_key.
-      // The prior split fetched baselined via .in('id', ids), matching the UUID `id` column against sd_key
-      // strings => 0 rows => baselined candidates dropped from the tier pool => belt_claimable_at_my_tier
-      // under-counted to 0. One fetch keyed by sd_key is correct for both kinds.
-      const allKeys = ranked.map((x) => x.key).filter(Boolean);
-      const cols = 'sd_key,id,sd_type,status,description,title,metadata,target_application,claiming_session_id';
-      let pool = [];
-      if (allKeys.length) {
-        const { data } = await sb.from('strategic_directives_v2').select(cols).in('sd_key', allKeys);
-        // QF-20260629-047: drop SDs already claimed by ANOTHER session — they are not claimable-to-me, so
-        // counting them inflates belt_claimable_at_my_tier and suppresses the tier-deficit idle message
-        // (which only fires at 0). Mirrors the forecaster's `if (d.claiming_session_id) continue;`. Keep
-        // rows claimed by THIS session (resume) and unclaimed rows.
-        pool = (data || []).filter((r) => !r.claiming_session_id || r.claiming_session_id === sessionId);
-      }
-      base.belt_claimable_at_my_tier = claimableForTier(pool, {
-        workerTierRank: tierCtx.worker_tier_rank,
-        tieringActive: tierCtx.tiering_active === true,
-      }).length;
-    } catch { /* fail-open: keep the agnostic count */ }
-    for (const x of ranked) {
-      if (x.kind === 'baselined') {
-        // SD-FDBK-FIX-WORKER-SELF-CLAIM-001: skip dependency-blocked SDs and orchestrator PARENTS
-        // (the view surfaces both; claim_sd enforces neither). Mirrors the draft-tier guard.
-        // SD-LEO-INFRA-WORKER-CLAIM-TIME-001 (FR-2): {cwd} adds the claim-time fitness axes so a
-        // baselined candidate unfit for THIS checkout is skipped before claiming.
-        if (!(await baselinedCandidateEligible(sb, x.key, { cwd: process.cwd(), ...tierCtx }))) continue;
-        if (await isSdInFlight(sb, x.key, sessionId)) continue;  // dedup: started or live-foreign-held
-        const claimed = await tryClaim(sb, x.key, sessionId, x.track);
-        if (claimed.ok) {
-          return { ...base, action: 'self_claimed', sd: x.key, track: x.track,
-            message: `Self-claimed ${x.key} from sd:next. Run: node scripts/sd-start.js ${x.key}. ${antiWinddownDirective(ranked.length)}` };
-        }
-      } else {
-        const result = await tryClaimDraftCandidate(sb, sessionId, base, x.row, tierCtx);
-        if (result) return result;
-      }
-    }
-  } catch { /* fail-open */ }
-
-  // 6.5 self-claim an open quick_fix. v_sd_next_candidates is SD-only, so open
-  // QFs are sourced here — strictly BELOW SD candidates and ABOVE idle, so a
-  // worker pulls an open QF instead of idling, but SD work always wins.
-  // SD-LEO-INFRA-MAKE-OPEN-QFS-001.
-  const qfClaimed = await selfClaimQuickFix(sb, sessionId, base);
-  if (qfClaimed) return qfClaimed;
-
-  // 7. idle -> recommend a wakeup (ScheduleWakeup is a HARNESS tool, not Node-callable)
-  // SD-LEO-INFRA-BELT-TIER-AWARE-CLAIMABILITY-001 (FR-2): when the ranked pool is non-empty but NONE of
-  // it is claimable at this worker's rung, say so explicitly — otherwise the agnostic count reads as
-  // "work exists for me" and the idle looks like a bug rather than a tier deficit.
-  const rankedAgnostic = base.belt_ranked_claimable ?? 0;
-  const claimableAtTier = base.belt_claimable_at_my_tier ?? rankedAgnostic;
-  // QF-20260630-761: only blame TIER when tiering is actually active. When tiering is OFF (degrade-to-1,
-  // <2 live workers) the tier axis is inert, so a 0-claimable belt with ranked>0 means the ranked items
-  // are ineligible for NON-tier reasons (orchestrator parents / clone build-trees / human-action / held)
-  // — a higher-tier worker could not take them either. Attributing it to "your rung" misdirects.
-  const tierNote = (rankedAgnostic > 0 && claimableAtTier === 0)
-    ? (base.belt_tiering_active === true
-        ? ` (${rankedAgnostic} ranked, but 0 claimable at your tier — all above your rung; a higher-tier worker must take them.)`
-        : ` (${rankedAgnostic} ranked, but 0 claimable by any worker — they are orchestrator parents / clone build-trees / human-action / held, not tier-blocked.)`)
-    : '';
-  // SD-LEO-INFRA-ADAPTIVE-COMMS-CADENCE-SHARED-PROTOCOL-001 (FR-6): opt-in tightening. A worker
-  // idling on the belt but awaiting a reply on a live comms thread (e.g. a blocked-item question
-  // to the coordinator) shouldn't wait a full baseline interval to notice the reply. ADDITIVE
-  // ONLY — never loosens the existing belt-driven recommendation, never touches claim/heartbeat.
-  // Fail-open: any error here falls through to the unchanged baseline recommendation.
-  let idleWakeupSeconds = DEFAULT_IDLE_WAKEUP_SECONDS;
-  let adaptiveCadenceNote = '';
-  try {
-    const signals = await getCommsActivitySignals(sb, sessionId);
-    const cadence = computeAdaptiveCadence(signals);
-    if (cadence.tight && cadence.intervalMs / 1000 < idleWakeupSeconds) {
-      idleWakeupSeconds = Math.round(cadence.intervalMs / 1000);
-      adaptiveCadenceNote = ` (tightened to ${idleWakeupSeconds}s — live comms thread: ${cadence.reason})`;
-    }
-  } catch { /* fail-open: keep the baseline recommendation */ }
-
-  return {
-    ...base,
-    action: 'idle',
-    recommended_wakeup_seconds: idleWakeupSeconds,
-    message: `No assignment and nothing claimable. IDLE.${tierNote} The /checkin skill must now call ScheduleWakeup(~${idleWakeupSeconds}s)${adaptiveCadenceNote} and proceed — never wait on a human.`,
+  const ctx = {
+    sb,
+    sessionId,
+    opts: { cliModel, cliEffort },
+    coordinatorId,
+    callsign,
+    mySd,
+    sessionRole,
+    sessionMetadata,
+    base: null,
+    helpers: CHECKIN_HELPERS,
   };
+  return runSteps(CHECKIN_STEPS, ctx);
 }
 
 /**
@@ -1929,6 +1377,12 @@ async function main() {
   const result = await runCheckin(sb, sessionId, { model: cliModel, effort: cliEffort });
   console.log(JSON.stringify(result, null, 2));
 }
+
+// SD-ARCH-HOTSPOT-CHECKIN-001: dependency injection for the lib/checkin/steps/* pipeline.
+// Steps destructure what they need from ctx.helpers instead of require()ing this file (which
+// would be circular). Every name below is either defined above in this file or imported at the
+// top (imports are referenced directly — never re-derived).
+const CHECKIN_HELPERS = { ws, tryClaim, ackMessage, extractSdFromAssignment, extractDirectedSd, isInformationalNudge, classifyDispatchIneligibility, registerRollCall, rehydrateCallsign, selfClearQuarantine, mergeCheckinModelEffort, recoverStrandedFinal, adoptOrphanInProgress, isSelfClaimDisabled, isGlobalStandDownActive, isBuildForbiddenSession, ensureActiveBaseline, isCriticalQfJumpEligible, tryClaimDraftCandidate, baselinedCandidateEligible, isSdInFlight, selfClaimQuickFix, selfHealStaleClaim, confirmRowGone, surfaceCoordinatorMessages, fetchDraftCandidates, fetchNewestDraftCandidates, fetchFleetCriticalCandidates, fetchRankedCandidates, sortByDispatchRank, resolveWorkerTierRank, isTieringActive, fetchLowerTierBacklogData, ladderTopRank, fetchFableWindowActive, claimableForTier, getCommsActivitySignals, computeAdaptiveCadence, antiWinddownDirective, ASSIGNMENT_RECENCY_WINDOW_MS, TERMINAL_CLAIM_ERRORS, QF_CANDIDATE_LIMIT, SELF_CLAIM_CANDIDATE_LIMIT, DEFAULT_IDLE_WAKEUP_SECONDS };
 
 module.exports = { extractSdFromAssignment, extractDirectedSd, isInformationalNudge, tryClaim, registerRollCall, ackMessage, isCoordinatorPush, surfaceCoordinatorMessages, rehydrateCallsign, runCheckin, resolveCheckin, assignFleetIdentityAtCheckin, selfClaimQuickFix, isAutoStartableQF, sortQfCandidatesBySeverity, QF_SEVERITY_RANK, isCriticalQfJumpEligible, CRITICAL_QF_JUMP_GRACE_MS, selfClaimDraftSd, fetchDraftCandidates, fetchNewestDraftCandidates, fetchFleetCriticalCandidates, fetchRankedCandidates, tryClaimDraftCandidate, draftDepsSatisfied, baselinedCandidateEligible, recoverStrandedFinal, adoptOrphanInProgress, isSelfClaimDisabled, isQuarantined, isParked, selfClearQuarantine, isGlobalStandDownActive, isSdInFlight, isForeignSessionLive, foreignClaimantBlocksSteal, selfHealStaleClaim, confirmRowGone, orderByRankMap, orderByFleetCriticalThenRank, sortByDispatchRank, DISPATCH_RANK_TTL_MS, PRIORITY_RANK, SD_KEY_RE, DEFAULT_IDLE_WAKEUP_SECONDS, STALE_QF_DAYS, antiWinddownDirective, mergeCheckinModelEffort, parseCheckinArgs };
